move tests to migrations tests
